remove jnr-ffi
