+++ tsec-libsodium/src/main/scala/tsec/internal/Chacha20Poly1305IETFConstants.scala
@@ -1,0 +1,13 @@
+package tsec.internal
+
+trait Chacha20Poly1305IETFConstants {
+
+  val crypto_aead_chacha20poly1305_ietf_KEYBYTES = 32
+
+  val crypto_aead_chacha20poly1305_ietf_NSECBYTES = 0
+
+  val crypto_aead_chacha20poly1305_ietf_NPUBBYTES = 12
+
+  val crypto_aead_chacha20poly1305_ietf_ABYTES = 16
+
+}
