+++ tsec-libsodium/src/main/scala/tsec/internal/HmacSha256Constants.scala
@@ -1,0 +1,8 @@
+package tsec.internal
+
+private[tsec] trait HmacSha256Constants {
+
+  val crypto_auth_hmacsha256_BYTES = 32
+
+  val crypto_auth_hmacsha256_KEYBYTES = 32
+}
