+++ tsec-libsodium/src/main/scala/tsec/internal/HmacSha512256Constants.scala
@@ -1,0 +1,8 @@
+package tsec.internal
+
+private[tsec] trait HmacSha512256Constants {
+
+  val crypto_auth_hmacsha512256_BYTES = 32L
+
+  val crypto_auth_hmacsha512256_KEYBYTES = 32L
+}
