+++ tsec-libsodium/src/main/scala/tsec/internal/HmacSha512Constants.scala
@@ -1,0 +1,8 @@
+package tsec.internal
+
+private[tsec] trait HmacSha512Constants {
+
+  val crypto_auth_hmacsha512_BYTES = 64L
+
+  val crypto_auth_hmacsha512_KEYBYTES = 32L
+}
