+++ tsec-libsodium/src/main/scala/tsec/internal/OriginalChacha20Poly1305Constants.scala
@@ -1,0 +1,11 @@
+package tsec.internal
+
+trait OriginalChacha20Poly1305Constants {
+
+  val crypto_aead_chacha20poly1305_KEYBYTES = 32
+
+  val crypto_aead_chacha20poly1305_NPUBBYTES = 8
+
+  val crypto_aead_chacha20poly1305_ABYTES = 16
+
+}
