+++ tsec-libsodium/src/main/scala/tsec/internal/ShortHashConstants.scala
@@ -1,0 +1,13 @@
+package tsec.internal
+
+private[tsec] trait ShortHashConstants {
+
+  val crypto_shorthash_siphash24_BYTES = 8
+
+  val crypto_shorthash_siphash24_KEYBYTES = 16
+
+  val crypto_shorthash_siphashx24_BYTES = 16
+
+  val crypto_shorthash_siphashx24_KEYBYTES = 16
+
+}
